[Docs] 메인 파트 폴더명을 변경하였습니다.
